Refactor analyze_docker_image
Created analyze.py which contains all the functions
that analyze_docker_image did.
The analyze_docker_image function has
been refactored and divided into multiple functions

Resolves https://github.com/vmware/tern/issues/212
